move annotation processor to core/annotations module, in order to reduce the extra dependency. add compiler arg to stop process execution in kernel
