+++ core/src/main/java/org/wso2/carbon/kernel/annotations/processor/ConfigurationProcessor.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.wso2.carbon.processors.annotation;
+package org.wso2.carbon.kernel.annotations.processor;
 
 
 import org.wso2.carbon.kernel.annotations.Configuration;
