+++ core/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,0 +1,1 @@
+org.wso2.carbon.kernel.annotations.processor.ConfigurationProcessor
