+++ pom.xml
@@ -26,10 +26,6 @@
     <artifactId>carbon-kernel</artifactId>
     <packaging>pom</packaging>
     <version>5.2.0-SNAPSHOT</version>
-    <modules>
-        <module>config-creator</module>
-        <module>processors</module>
-    </modules>
     <name>WSO2 Carbon Kernel - Aggregate Pom</name>
     <description>WSO2 Carbon Kernel Aggregate Pom</description>
     <url>http://wso2.com</url>
@@ -50,7 +46,6 @@
                 <module>launcher</module>
                 <module>core</module>
                 <module>plugins</module>
-                <module>processors</module>
                 <module>tools</module>
                 <module>config-creator</module>
                 <module>pax-exam-container-carbon</module>
