--- processors/annotation-processor/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,0 @@
-org.wso2.carbon.processors.annotation.ConfigurationProcessor
